renamed a file and changed json file
