+++ examples/autoshutoff-jupyter-kernel/.saturn/saturn.json
@@ -2,7 +2,7 @@
   "name": "example-autoshutoff-jupyter-kernel",
   "image_uri": "saturncloud/saturn:2021.09.20",
   "description": "POC of using Jupyter kernel information for shutting off a resource",
-  "working_directory": "/home/jovyan/git-repos/autoshutoff",
+  "working_directory": "/home/jovyan/git-repos/examples/examples/autoshutoff-jupyter-kernel",
   "environment_variables": {
     "SATURN_APP_URL": "https://app.internal.saturnenterprise.io/"
   },
