+++ examples/autoshutoff-jupyter-kernel/userautoshutoff.py
